"Refactor" Game Screen class
This is not a real refactoring, but moves most of the Scene2D UI code in Game
Screen class to a new file.
The unreadable blocks of code still exist, and I should refactor them later.
However, refactoring is too boring, and I don't want to do it for a (not too)
long time, unless I consider it to be mandatory to continue.